+++ src/main/java/com/hackathon/churn/ChurnGraphqlApiApplication.java
@@ -2,8 +2,10 @@ package com.hackathon.churn;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.data.mongodb.repository.config.EnableMongoRepositories;
 
 @SpringBootApplication
+@EnableMongoRepositories(basePackages = "com.hackathon.churn.Repository")
 public class ChurnGraphqlApiApplication {
 
 	public static void main(String[] args) {