+++ src/main/java/com/hackathon/churn/controller/UsuarioController.java
@@ -1,0 +1,41 @@
+package com.hackathon.churn.controller;
+
+import jakarta.validation.Valid;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.ResponseEntity;
+import org.springframework.security.crypto.password.PasswordEncoder;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+import com.hackathon.churn.Repository.UsuarioRepository;
+import com.hackathon.churn.usuario.DadosUsuario;
+import com.hackathon.churn.usuario.Usuario;
+import org.springframework.web.util.UriComponentsBuilder;
+
+@RestController
+@RequestMapping("/usuarios")
+public class UsuarioController {
+
+    @Autowired
+    private UsuarioRepository repository;
+
+    @Autowired
+    private PasswordEncoder passwordEncoder;
+
+    @PostMapping
+    public ResponseEntity cadastrar(@RequestBody @Valid DadosUsuario dados, UriComponentsBuilder uriBuilder) {
+        // Verifica se usuário já existe
+        if (repository.findByLogin(dados.login()) != null) {
+            return ResponseEntity.badRequest().body("Usuário já existe");
+        }
+
+        var usuario = new Usuario();
+        usuario.setLogin(dados.login());
+        usuario.setSenha(passwordEncoder.encode(dados.senha()));
+        repository.save(usuario);
+
+        var uri = uriBuilder.path("/usuarios/{id}").buildAndExpand(usuario.getId()).toUri();
+        return ResponseEntity.created(uri).body("Usuário criado com sucesso"); // Retorna 201 Created
+    }
+}