+++ src/main/java/com/hackathon/churn/security/SecurityConfiguration.java
@@ -22,27 +22,28 @@ public class SecurityConfiguration {
 
     @Bean
     public SecurityFilterChain securityFilterChain(HttpSecurity http) throws Exception {
-        return
-                http.csrf(csrf -> csrf.disable())
-                        .sessionManagement(sm -> sm.sessionCreationPolicy(SessionCreationPolicy.STATELESS))
-                        .authorizeHttpRequests(req -> {
-                            req.requestMatchers("/graphql").permitAll();
-                            req.requestMatchers("/login").permitAll();
-                            req.requestMatchers("/usuarios").permitAll();
-                            req.anyRequest().authenticated();
+        return http.csrf(csrf -> csrf.disable())
+                .sessionManagement(sm -> sm.sessionCreationPolicy(SessionCreationPolicy.STATELESS))
+                .authorizeHttpRequests(req -> {
+                    req.requestMatchers("/login").permitAll();
+                    req.requestMatchers("/usuarios").permitAll();
+                    req.requestMatchers("/graphiql").permitAll();
+                    req.requestMatchers("/graphql").permitAll(); // Liberado no Spring Security, validado via
+                                                                 // Token/Contexto
+                    req.anyRequest().authenticated();
 
-
-                        })
-                        .addFilterBefore(securityFilter, UsernamePasswordAuthenticationFilter.class)
-                        .build();
+                })
+                .addFilterBefore(securityFilter, UsernamePasswordAuthenticationFilter.class)
+                .build();
     }
+
     @Bean
     public AuthenticationManager manager(AuthenticationConfiguration configuration) throws Exception {
         return configuration.getAuthenticationManager();
     }
 
     @Bean
-    public PasswordEncoder passwordEncoder(){
+    public PasswordEncoder passwordEncoder() {
         return new BCryptPasswordEncoder();
     }